requirements.txt
tensorflow
matplotlib
psycopg2-binary
pytest
numpy
fastapi
uvicorn
sqlalchemy
opencv-python
python-dotenv
httpx
